Return Tasks from File, MVar, and Runtime kernel functions
Several kernel entry points returned raw scalars where Elm expects a Task, so the scheduler
misinterpreted their results. File.size and File.modificationTime now return Task Int (succeeding
with a boxed Int, failing with a string on fstat/stat error); MVar.new boxes the new id and returns
it via taskSucceed; and Runtime.random returns Task Float. Update the corresponding declarations in
File.hpp, Runtime.hpp, KernelExports.h, and the *Exports.cpp shims to match.

Apply an effect-manager init closure before spawning it in PlatformRuntime::setupEffects. Some
effect managers register `init` as a 0-arg closure that produces the init Task when applied rather
than as an evaluated Task; detect a Tag_Closure init, apply it to obtain the actual Task, and
otherwise hand a Closure to rawSpawn whose bytes would be misread as a Task header, silently
dropping the init effect.

Remove the ECO_SCHED_TRACE scaffolding and the stepProcess heap-tag check and unknown-ctor logging
that were added for debugging the scheduler.

+++ eco-kernel-cpp/src/eco/KernelExports.h
@@ -41,7 +41,7 @@ uint64_t Eco_Kernel_File_open(uint64_t path, uint64_t mode);
 uint64_t Eco_Kernel_File_close(uint64_t handle);
 
 // Get file size via handle. Returns Int (unboxed).
-int64_t Eco_Kernel_File_size(uint64_t handle);
+uint64_t Eco_Kernel_File_size(uint64_t handle);
 
 // Acquire file lock (blocks). Returns Task Never ().
 uint64_t Eco_Kernel_File_lock(uint64_t path);
@@ -62,7 +62,7 @@ uint64_t Eco_Kernel_File_findExecutable(uint64_t name);
 uint64_t Eco_Kernel_File_list(uint64_t path);
 
 // Get file modification time. Returns Int (milliseconds since epoch, unboxed).
-int64_t Eco_Kernel_File_modificationTime(uint64_t path);
+uint64_t Eco_Kernel_File_modificationTime(uint64_t path);
 
 // Get current working directory. Returns String (boxed).
 uint64_t Eco_Kernel_File_getCwd();
@@ -146,7 +146,10 @@ uint64_t Eco_Kernel_Process_wait(uint64_t handle);
 //===----------------------------------------------------------------------===//
 
 // Create new empty MVar. Returns Int (MVar id, unboxed).
-int64_t Eco_Kernel_MVar_new();
+// Returns a Task that succeeds with a freshly-allocated MVar id (boxed Int).
+// (Elm side: `Eco.Kernel.MVar.new : Task Never Int`, then `Task.map MVar` in
+// Eco.MVar wraps the Int in the MVar constructor.)
+uint64_t Eco_Kernel_MVar_new();
 
 // Read MVar (blocks until full). Returns value (boxed).
 // typeTag is unused at runtime but required by Elm type system.
@@ -169,7 +172,7 @@ uint64_t Eco_Kernel_MVar_drop(uint64_t id);
 uint64_t Eco_Kernel_Runtime_dirname();
 
 // Get random Float. Returns Float (unboxed).
-double Eco_Kernel_Runtime_random();
+uint64_t Eco_Kernel_Runtime_random();
 
 // Persist REPL state to runtime storage. Returns Task Never ().
 uint64_t Eco_Kernel_Runtime_saveState(uint64_t state);
